Code Uploaded
Code and graphs done in Jupyter Notebook. The report and presentation focuses on 2021-2022, however I also performed the same analysis for 2020-2021 and 2019-2020 to check if results were consistent. I didn't include these in the report as the population data was specifically recorded in 2021, so I was concerned that it may have changed and affected my results.